pyproject.toml
[project]
name = "matmaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opik>=1.8.71",
    "bohr-agent-sdk==0.1.106",
    "oss2>=2.18.0",
    "aiofiles>=24.1.0",
    "pymysql>=1.1.1",
    "anyio>=4.10.0",
    "pytest-asyncio>=1.1.0",
    "litellm>=1.76.1",
    "google-adk[a2a,eval,extensions,test]==1.16.0",
    "pre-commit>=4.3.0",
    "bohrium-sdk>=0.14.0",
    "toolsy>=0.1.2",
    "deepmerge>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents"]
